feat(morgan): create morgan logger
