+++ src/server.ts
@@ -3,12 +3,11 @@ import helmet from 'helmet';
 import cors from 'cors';
 import routes from './routes';
 import bodyParser from 'body-parser';
+import cookieParser from 'cookie-parser';
 import morgan from 'morgan';
-import fs from 'fs';
-import path from 'path';
-import logger from './utils/logger';
-import { logMiddleware } from './utils/logMiddleware';
+import { logMiddleware } from './middleware/logMiddleware';
 import { dbCreateConnection } from './utils/db';
+import { createStream } from './utils/morganLogger';
 
 const bootstrap = async () => {
   const app: Application = express();
@@ -17,34 +16,17 @@ const bootstrap = async () => {
   app.use(helmet());
   app.use(bodyParser.json());
   app.use(bodyParser.urlencoded({ extended: false }));
+  app.use(cookieParser());
   app.use(logMiddleware());
+  app.use(morgan('combined', { stream: createStream() }));
+
+  app.use(morgan('dev'));
+
+  await dbCreateConnection();
   app.use('/', routes);
 
-  try {
-    const date = new Date();
-    const year = date.getFullYear().toString();
-
-    let month = (date.getMonth() + 1).toString();
-    month = +month < 10 ? '0' + month : month;
-
-    let day = date.getDate().toString();
-    day = +day < 10 ? '0' + day : day;
-
-    const accessLogStream = fs.createWriteStream(
-      path.join(__dirname, `../logs/${year}-${month}-${day}.log`),
-      {
-        flags: 'a',
-      },
-    );
-    app.use(morgan('combined', { stream: accessLogStream }));
-  } catch (err) {
-    console.log(err);
-  }
-
-  dbCreateConnection();
-
   app.listen(4120, () => {
-    logger.info('✅  Server listening on port: 4120');
+    console.log('✅  Server listening on port: 4120');
   });
 };
 bootstrap();
