+++ src/utils/db.ts
@@ -28,9 +28,8 @@ export const dbCreateConnection = async (): Promise<Connection | null> => {
   try {
     const conn = await createConnection(config);
     console.log(
-      `Database connection success. Connection name: '${conn.name}' Database: '${conn.options.database}'`,
+      `✅  Database connection success. Connection name: '${conn.name}' Database: '${conn.options.database}'`,
     );
-    console.log(__dirname + '/../shared/entities/*.{js,ts}');
   } catch (err: any) {
     if (err.name === 'AlreadyHasActiveConnectionError') {
       const activeConnection = getConnectionManager().get(config.name);
